Refactorize root cmd
