+++ app/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/giansalex/binance-stoploss/app/cmd"
+
+func main() {
+	cmd.Execute()
+}
